+++ HarpGuru/OverlayMenu/Menus/CovariantMenu/components/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import {MenuContainer} from '../../MenuContainer'
 import type {MenuProps} from '../../../types'
-import { DisplayModeOption, HarpKeyOption, RootPitchOption, PozitionOption } from '../../../MenuOptions'
+import { DisplayModeOption, HarpKeyOption, RootPitchOption, PozitionOption } from '../../../Options'
 
 
 export const CovariantMenu = (props: MenuProps): React.ReactElement => {
