+++ HarpGuru/OverlayMenu/Menus/LayoutMenu/components/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { MenuContainer} from '../../MenuContainer'
 import type { MenuProps } from '../../../types'
-import { ApparatusOption, DisplayModeOption } from '../../../MenuOptions'
+import { ApparatusOption, DisplayModeOption } from '../../../Options'
 
 
 export const LayoutMenu = (props: MenuProps): React.ReactElement => {
